moved allureStep method implementation from framework to report folder
